% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_fake_phone_numbers.R
\name{detect_fake_phone_numbers}
\alias{detect_fake_phone_numbers}
\title{Detecting fake mobile numbers in Iran}
\usage{
detect_fake_phone_numbers(phone_numbers)
}
\arguments{
\item{phone_numbers}{}
}
\value{
TRUE or FALSE
}
\description{
Detecting fake mobile numbers in Iran
}
\examples{
detect_fake_phone_numbers("09121234567")
}
