% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsi_keyboard.R
\name{farsi_keyboard}
\alias{farsi_keyboard}
\title{Correcting characters from Arabic keyboards}
\usage{
farsi_keyboard(dataset, coln)
}
\arguments{
\item{dataset}{Your dataset}

\item{coln}{The column you want to clean}
}
\value{
Dataset with Persian characters corrected
}
\description{
Correcting characters from Arabic keyboards
}
