% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_state_en.R
\name{mutate_state_en}
\alias{mutate_state_en}
\title{Ostan to Province}
\usage{
mutate_state_en(data, column_name)
}
\arguments{
\item{data}{Pass the data you are working on}

\item{column_name}{the is the column that contains Persian names of Ostan(استان)}
}
\value{
state_en columns that contains English name of states in Iran.
}
\description{
This function gets Persian state name and transform the Persian col to English which is standard names for Iran's states in English
}
