% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psave_plot.R
\name{psave_plot}
\alias{psave_plot}
\title{Save plots easily}
\usage{
psave_plot(
  plot,
  height = 8,
  width = 12,
  dpi = 300,
  units = "in",
  path = getwd()
)
}
\arguments{
\item{path}{}
}
\description{
This function save your plots in ggplot2 in a dir named plot that may or may not exists in you wd dir.
}
