% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasmio_theme.R
\name{rasmio_theme}
\alias{rasmio_theme}
\title{Rasmio Theme}
\usage{
rasmio_theme(
  persian_font = "Sahel FD",
  title_size = 28,
  subtitle_size = 22,
  legend_position = "bottom",
  axis_size = 18
)
}
\arguments{
\item{persian_font}{font family. Defaults tp Sahel Family. Make sure you have it!}

\item{title_size}{size}

\item{subtitle_size}{description}

\item{legend_position}{legend position}

\item{axis_size}{this value controls the font size of the axis for both x and y}
}
\value{
Nice Theme for plots in Persian!
}
\description{
This function handles theme with Rasmio standards
}
