% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_map_fa.R
\name{theme_map_fa}
\alias{theme_map_fa}
\title{Theme map Farsi}
\usage{
theme_map_fa(
  persian_font = "Sahel FD",
  title_size = 32,
  subtitle_size = 22,
  legend_position = "bottom",
  facet_text_size = 16
)
}
\arguments{
\item{persian_font}{font}

\item{title_size}{size}

\item{subtitle_size}{description}

\item{legend_position}{legend position}
}
\value{
Nice looking Persian Map Theme in ggplot2
}
\description{
This function handles theme in Persian plot
}
