% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_en_number.R
\name{to_en_numbers}
\alias{to_en_numbers}
\title{Persian numbers to English Numbers}
\usage{
to_en_numbers(x)
}
\arguments{
\item{x}{Persian number}
}
\value{
English Number
}
\description{
If you have a number in Persian, this functions turns that number to English number.
}
\examples{
to_en_numbers("123")
}
