% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_fa_numbers.R
\name{to_fa_numbers}
\alias{to_fa_numbers}
\title{English numbers to Persian numbers}
\usage{
to_fa_numbers(x)
}
\arguments{
\item{x}{The English number}
}
\value{
Persian Number
}
\description{
English numbers to Persian numbers
}
\examples{
to_fa_numbers(123)
}
