% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/churn.R
\name{calculate_churn_rate}
\alias{calculate_churn_rate}
\title{churn rates by year}
\usage{
calculate_churn_rate(data, year, product_id = "product_id")
}
\arguments{
\item{data}{Inpute your data}

\item{year}{The years you are calculating churn rate}

\item{product_id}{the unique id}
}
\value{
churn rate
}
\description{
This function returns churn rate for product id in year.
}
