% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/farsi_keyboard.R
\name{farsi_keyboard}
\alias{farsi_keyboard}
\title{correcting charachters from Arabic keyboards}
\usage{
farsi_keyboard(dataset, coln)
}
\arguments{
\item{dataset}{your dataset}

\item{coln}{the column you want to clean}
}
\value{
Persian charachters
}
\description{
correcting charachters from Arabic keyboards
}
